xml rpc
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@2908 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ BioradHeader.h
@@ -21,8 +21,8 @@
 
 struct BioradHeader{
 	short int nx,ny,npic;
-	long int unused0;
-	long int notesAvailable;
+	 int unused0;
+	 int notesAvailable;
 	short int byte_format;
 	char unused1[38];
 	short int magicNumber;
